Bug Fixes and Game Staged
Bugs fixed and game has been staged. Possible things to look for in the future if i return is a more complicated AI for the ghosts as its very simple at the moment and flawed. As well as upping the difficulty level as you go although its fairyl difficult at the moment as it is.